Added bias.py.
